server/requirements.txt
colorama==0.4.6
pycryptodome==3.17
PyYAML==6.0
requests==2.28.2
rsa==4.9
